feat(web): middleware implementation
I thought adding middleware was going to be difficult but turns out, it
is pretty easy.
Now you can easily define a middleware using the `WebServer::middleware`
function, and your middlewares will be executed in the exact order as
you define them.

Other Changes:
- while adding the middlewares, it occured to me that I can simplify the
routing code by removing the `RouteHandler` struct and replacing it with
a `Box` wrapping a closure function for registering routes
- clear up and update the documentation here and there